moved changes back into _read_rows.py
