+++ google/cloud/bigtable/data/_async/_read_rows.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Google LLC
+
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,40 +12,35 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
+
 from __future__ import annotations
 
 from typing import (
-    List,
     Any,
-    AsyncIterable,
-    AsyncIterator,
-    AsyncGenerator,
-    Iterator,
-    Callable,
-    Awaitable,
 )
-import sys
-import time
-import asyncio
-from functools import partial
-from grpc.aio import RpcContext
 
-from google.cloud.bigtable_v2.types.bigtable import ReadRowsResponse
-from google.cloud.bigtable_v2.services.bigtable.async_client import BigtableAsyncClient
-from google.cloud.bigtable.data.row import Row, _LastScannedRow, Cell
+from google.cloud.bigtable_v2.types import ReadRowsRequest as ReadRowsRequestPB
+from google.cloud.bigtable_v2.types import RowSet as RowSetPB
+from google.cloud.bigtable_v2.types import RowRange as RowRangePB
+
+from google.cloud.bigtable.data.row import Row, Cell
 from google.cloud.bigtable.data.exceptions import InvalidChunk
 from google.cloud.bigtable.data.exceptions import _RowSetComplete
-from google.cloud.bigtable.data.exceptions import IdleTimeout
+from google.cloud.bigtable.data.exceptions import RetryExceptionGroup
+from google.cloud.bigtable.data._helpers import _attempt_timeout_generator
+from google.cloud.bigtable.data._helpers import _convert_retry_deadline
+from google.cloud.bigtable.data._helpers import _make_metadata
+
 from google.api_core import retry_async as retries
 from google.api_core.retry_streaming_async import AsyncRetryableGenerator
 from google.api_core.retry import exponential_sleep_generator
 from google.api_core import exceptions as core_exceptions
-from google.cloud.bigtable.data._helpers import _make_metadata
-from google.cloud.bigtable.data._helpers import _attempt_timeout_generator
-from google.cloud.bigtable.data._helpers import _convert_retry_deadline
+
+class _ResetRow(Exception):
+    pass
 
 
-class _ReadRowsOperationAsync(AsyncIterable[Row]):
+class _ReadRowsOperationAsync():
     """
     ReadRowsOperation handles the logic of merging chunks from a ReadRowsResponse stream
     into a stream of Row objects.
@@ -60,54 +55,67 @@ class _ReadRowsOperationAsync(AsyncIterable[Row]):
 
     def __init__(
         self,
-        request: dict[str, Any],
-        client: BigtableAsyncClient,
-        *,
-        operation_timeout: float = 600.0,
-        attempt_timeout: float | None = None,
+        query,
+        table,
+        operation_timeout: float,
+        attempt_timeout: float,
     ):
-        """
-        Args:
-          - request: the request dict to send to the Bigtable API
-          - client: the Bigtable client to use to make the request
-          - operation_timeout: the timeout to use for the entire operation, in seconds
-          - attempt_timeout: the timeout to use when waiting for each individual grpc request, in seconds
-                If not specified, defaults to operation_timeout
-        """
-        self._last_emitted_row_key: bytes | None = None
-        self._emit_count = 0
-        self._request = request
-        self.operation_timeout = operation_timeout
-        # use generator to lower per-attempt timeout as we approach operation_timeout deadline
-        attempt_timeout_gen = _attempt_timeout_generator(
+        self.attempt_timeout_gen = _attempt_timeout_generator(
             attempt_timeout, operation_timeout
         )
-        row_limit = request.get("rows_limit", 0)
-        # lock in paramters for retryable wrapper
-        retryable_stream = partial(
-            self._read_rows_retryable_attempt,
-            client.read_rows,
-            attempt_timeout_gen,
-            row_limit,
+        self.operation_timeout = operation_timeout
+        query_dict = query._to_dict() if not isinstance(query, dict) else query
+        self.request = ReadRowsRequestPB(**query_dict, table_name=table.table_name)
+        if table.app_profile_id:
+            self.request.app_profile_id = table.app_profile_id
+        self.table = table
+        self._last_yielded_row_key = None
+        self._remaining_count = self.request.rows_limit or None
+        self._metadata = _make_metadata(
+            table.table_name,
+            table.app_profile_id,
         )
-        retryable_as_list = partial(
-            self._read_rows_retryable_attempt_as_list,
-            client.read_rows,
-            attempt_timeout_gen,
-            row_limit,
+
+    async def make_retry_stream(self):
+        stream, transient_errors = self._make_stream_helper(AsyncRetryableGenerator, self.read_rows_attempt)
+        try:
+            async for row in stream:
+                yield row
+        except core_exceptions.RetryError:
+            self._raise_retry_error(transient_errors)
+
+    async def make_retry_stream_as_list(self):
+        retry_fn, transient_errors = self._make_stream_helper(retries.retry_target, self.read_rows_attempt_as_list)
+        try:
+            return await retry_fn
+        except core_exceptions.RetryError:
+            self._raise_retry_error(transient_errors)
+
+    def _raise_retry_error(self, transient_errors):
+        timeout_value = self.operation_timeout
+        timeout_str = f" of {timeout_value:.1f}s" if timeout_value is not None else ""
+        error_str = f"operation_timeout{timeout_str} exceeded"
+        new_exc = core_exceptions.DeadlineExceeded(
+            error_str,
         )
+        source_exc = None
+        if transient_errors:
+            source_exc = RetryExceptionGroup(transient_errors)
+        new_exc.__cause__ = source_exc
+        raise new_exc from source_exc
+
+    def _make_stream_helper(self, retry_wrapper, attempt_fn):
+        transient_errors = []
         predicate = retries.if_exception_type(
             core_exceptions.DeadlineExceeded,
             core_exceptions.ServiceUnavailable,
             core_exceptions.Aborted,
         )
-
         def on_error_fn(exc):
             if predicate(exc):
-                self.transient_errors.append(exc)
-
-        self._stream: AsyncGenerator[Row, None] | None = AsyncRetryableGenerator(
-            retryable_stream,
+                transient_errors.append(exc)
+        retry_fn = retry_wrapper(
+            attempt_fn,
             predicate,
             exponential_sleep_generator(
                 0.01, 60, multiplier=2
@@ -115,124 +123,32 @@ class _ReadRowsOperationAsync(AsyncIterable[Row]):
             self.operation_timeout,
             on_error_fn,
         )
-        # self._as_list_fn = retries.retry_target(
-        #     retryable_as_list,
-        #     predicate,
-        #     exponential_sleep_generator(
-        #         0.01, 60, multiplier=2
-        #     ),
-        #     self.operation_timeout,
-        #     on_error_fn,
-        # )
-        self._as_list_fn = retryable_as_list
-        # contains the list of errors that were retried
-        self.transient_errors: List[Exception] = []
+        return retry_fn, transient_errors
 
-    def __aiter__(self) -> AsyncIterator[Row]:
-        """Implements the AsyncIterable interface"""
-        return self
-
-    async def __anext__(self) -> Row:
-        """Implements the AsyncIterator interface"""
-        if self._stream is not None:
-            return await self._stream.__anext__()
-        else:
-            raise asyncio.InvalidStateError("stream is closed")
-
-    async def aclose(self):
-        """Close the stream and release resources"""
-        if self._stream is not None:
-            await self._stream.aclose()
-        self._stream = None
-        self._emitted_seen_row_key = None
-
-    async def _read_rows_retryable_attempt_as_list(
-        self,
-        gapic_fn: Callable[..., Awaitable[AsyncIterable[ReadRowsResponse]]],
-        timeout_generator: Iterator[float],
-        total_row_limit: int,
-    ) -> List[Row]:
-        output_list = [row async for row in self._read_rows_retryable_attempt(gapic_fn, timeout_generator, total_row_limit)]
-        return output_list
-
-    async def _read_rows_retryable_attempt(
-        self,
-        gapic_fn: Callable[..., Awaitable[AsyncIterable[ReadRowsResponse]]],
-        timeout_generator: Iterator[float],
-        total_row_limit: int,
-    ) -> AsyncGenerator[Row, None]:
-        """
-        Retryable wrapper for merge_rows. This function is called each time
-        a retry is attempted.
-
-        Some fresh state is created on each retry:
-          - grpc network stream
-          - state machine to hold merge chunks received from stream
-        Some state is shared between retries:
-          - _last_emitted_row_key is used to ensure that
-            duplicate rows are not emitted
-          - request is stored and (potentially) modified on each retry
-        """
-        if self._last_emitted_row_key is not None:
+    async def read_rows_attempt(self):
+        if self._last_yielded_row_key is not None:
             # if this is a retry, try to trim down the request to avoid ones we've already processed
             try:
-                self._request["rows"] = _ReadRowsOperationAsync._revise_request_rowset(
-                    row_set=self._request.get("rows", None),
-                    last_seen_row_key=self._last_emitted_row_key,
+                self.request.rows = self._revise_request_rowset(
+                    row_set=self.request.rows,
+                    last_seen_row_key=self._last_yielded_row_key,
                 )
             except _RowSetComplete:
-                # if there are no rows left to process, we're done
-                # This is not expected to happen often, but could occur if
-                # a retry is triggered quickly after the last row is emitted
                 return
-            # revise next request's row limit based on number emitted
-            if total_row_limit:
-                new_limit = total_row_limit - self._emit_count
-                if new_limit == 0:
-                    # we have hit the row limit, so we're done
-                    return
-                elif new_limit < 0:
-                    raise RuntimeError("unexpected state: emit count exceeds row limit")
-                else:
-                    self._request["rows_limit"] = new_limit
-        metadata = _make_metadata(
-            self._request.get("table_name", None),
-            self._request.get("app_profile_id", None),
-        )
-        new_gapic_stream: RpcContext = await gapic_fn(
-            self._request,
-            timeout=next(timeout_generator),
-            metadata=metadata,
-        )
-        try:
-            stream = _ReadRowsOperationAsync.merge_row_response_stream(
-                new_gapic_stream
-            )
-            # run until we get a timeout or the stream is exhausted
-            async for new_item in stream:
-                # if (
-                #     self._last_emitted_row_key is not None
-                #     and new_item.row_key <= self._last_emitted_row_key
-                # ):
-                #     raise InvalidChunk("Last emitted row key out of order")
-                # don't yeild _LastScannedRow markers; they
-                # should only update last_seen_row_key
-                if not isinstance(new_item, _LastScannedRow):
-                    yield new_item
-                    self._emit_count += 1
-                self._last_emitted_row_key = new_item.row_key
-                if total_row_limit and self._emit_count >= total_row_limit:
-                    return
-        except (Exception, GeneratorExit) as exc:
-            # ensure grpc stream is closed
-            new_gapic_stream.cancel()
-            raise exc
+        self.request.rows_limit = self._remaining_count
+        s = await self.table.client._gapic_client.read_rows(self.request, timeout=next(self.attempt_timeout_gen), metadata=self._metadata)
+        s = self.chunk_stream(s)
+        return self.merge_rows(s)
+
+    async def read_rows_attempt_as_list(self):
+        stream = await self.read_rows_attempt()
+        return [row async for row in stream]
 
     @staticmethod
     def _revise_request_rowset(
-        row_set: dict[str, Any] | None,
+        row_set: RowSetPB,
         last_seen_row_key: bytes,
-    ) -> dict[str, Any]:
+    ) -> RowSetPB:
         """
         Revise the rows in the request to avoid ones we've already processed.
 
@@ -244,188 +160,142 @@ class _ReadRowsOperationAsync(AsyncIterable[Row]):
         """
         # if user is doing a whole table scan, start a new one with the last seen key
         if row_set is None or (
-            len(row_set.get("row_ranges", [])) == 0
-            and len(row_set.get("row_keys", [])) == 0
+            not row_set.row_ranges and not row_set.row_keys is None
         ):
             last_seen = last_seen_row_key
-            return {
-                "row_keys": [],
-                "row_ranges": [{"start_key_open": last_seen}],
-            }
+            return RowSetPB(row_ranges=[RowRangePB(start_key_open=last_seen)])
         # remove seen keys from user-specific key list
-        row_keys: list[bytes] = row_set.get("row_keys", [])
-        adjusted_keys = [k for k in row_keys if k > last_seen_row_key]
+        adjusted_keys : list[bytes] = [k for k in row_set.row_keys if k > last_seen_row_key]
         # adjust ranges to ignore keys before last seen
-        row_ranges: list[dict[str, Any]] = row_set.get("row_ranges", [])
-        adjusted_ranges = []
-        for row_range in row_ranges:
-            end_key = row_range.get("end_key_closed", None) or row_range.get(
-                "end_key_open", None
-            )
+        adjusted_ranges : list[RowRangePB] = []
+        for row_range in row_set.row_ranges:
+            end_key = row_range.end_key_closed or row_range.end_key_open
             if end_key is None or end_key > last_seen_row_key:
                 # end range is after last seen key
                 new_range = row_range.copy()
-                start_key = row_range.get("start_key_closed", None) or row_range.get(
-                    "start_key_open", None
-                )
+                start_key = row_range.start_key_closed or row_range.start_key_open
                 if start_key is None or start_key <= last_seen_row_key:
                     # replace start key with last seen
-                    new_range["start_key_open"] = last_seen_row_key
-                    new_range.pop("start_key_closed", None)
+                    new_range.start_key_open = last_seen_row_key
+                    new_range.start_key_closed = None
                 adjusted_ranges.append(new_range)
         if len(adjusted_keys) == 0 and len(adjusted_ranges) == 0:
             # if the query is empty after revision, raise an exception
             # this will avoid an unwanted full table scan
             raise _RowSetComplete()
-        return {"row_keys": adjusted_keys, "row_ranges": adjusted_ranges}
+        return RowSetPB(row_keys=adjusted_keys, row_ranges=adjusted_ranges)
 
     @staticmethod
-    async def merge_row_response_stream(
-        response_generator: AsyncIterable[ReadRowsResponse],
-    ) -> AsyncGenerator[Row, None]:
-        """
-        Consume chunks from a ReadRowsResponse stream into a set of Rows
+    async def chunk_stream(stream):
+        prev_key = None
 
-        Args:
-          - response_generator: AsyncIterable of ReadRowsResponse objects. Typically
-                this is a stream of chunks from the Bigtable API
-        Returns:
-            - AsyncGenerator of Rows
-        Raises:
-            - InvalidChunk: if the chunk stream is invalid
-        """
-        row = None
-        chunk_list = []
-        async for row_response in response_generator:
-            # unwrap protoplus object for increased performance
-            response_pb = row_response._pb
-            if len(response_pb.chunks) == 0:
-                # check for last scanned key
-                last_scanned = response_pb.last_scanned_row_key
-                # if the server sends a scan heartbeat, notify the state machine.
-                if last_scanned:
-                    yield _LastScannedRow(last_scanned)
-            else:
-                # process new chunks through the state machine.
-                for chunk in response_pb.chunks:
-                    chunk_list.append(chunk)
-                    if chunk.commit_row:
-                        yield Row._from_chunks(chunk_list)
-                        chunk_list.clear()
+        async for resp in stream:
+            resp = resp._pb
 
-        if row is not None:
-            # read rows is complete, but there's still data in the merger
-            raise InvalidChunk("read_rows completed with partial state remaining")
+            if resp.last_scanned_row_key:
+                if prev_key is not None and resp.last_scanned_row_key >= prev_key:
+                    raise InvalidChunk("last scanned out of order")
+                prev_key = resp.last_scanned_row_key
 
+            current_key = None
 
-class ReadRowsAsyncIterator(AsyncIterable[Row]):
-    """
-    Async iterator for ReadRows responses.
+            for c in resp.chunks:
+                if current_key is None:
+                    current_key = c.row_key
+                    if current_key is None:
+                        raise InvalidChunk("first chunk is missing a row key")
+                    elif prev_key and current_key <= prev_key:
+                        raise InvalidChunk("out of order row key")
 
-    Supports the AsyncIterator protocol for use in async for loops,
-    along with:
-      - `aclose` for closing the underlying stream
-      - `active` for checking if the iterator is still active
-      - an internal idle timer for closing the stream after a period of inactivity
-    """
+                yield c
 
-    def __init__(self, merger: _ReadRowsOperationAsync):
-        self._merger: _ReadRowsOperationAsync = merger
-        self._error: Exception | None = None
-        self._last_interaction_time = time.monotonic()
-        self._idle_timeout_task: asyncio.Task[None] | None = None
-        # wrap merger with a wrapper that properly formats exceptions
-        self._next_fn = _convert_retry_deadline(
-            self._merger.__anext__,
-            self._merger.operation_timeout,
-            self._merger.transient_errors,
-            is_async=True,
-        )
+                if c.reset_row:
+                    current_key = None
+                elif c.commit_row:
+                    prev_key = current_key
 
-    async def _start_idle_timer(self, idle_timeout: float):
-        """
-        Start a coroutine that will cancel a stream if no interaction
-        with the iterator occurs for the specified number of seconds.
+    async def merge_rows(self, chunks):
+        it = chunks.__aiter__()
 
-        Subsequent access to the iterator will raise an IdleTimeout exception.
-
-        Args:
-          - idle_timeout: number of seconds of inactivity before cancelling the stream
-        """
-        self._last_interaction_time = time.monotonic()
-        if self._idle_timeout_task is not None:
-            self._idle_timeout_task.cancel()
-        self._idle_timeout_task = asyncio.create_task(
-            self._idle_timeout_coroutine(idle_timeout)
-        )
-        if sys.version_info >= (3, 8):
-            self._idle_timeout_task.name = f"{self.__class__.__name__}.idle_timeout"
-
-    @property
-    def active(self):
-        """
-        Returns True if the iterator is still active and has not been closed
-        """
-        return self._error is None
-
-    async def _idle_timeout_coroutine(self, idle_timeout: float):
-        """
-        Coroutine that will cancel a stream if no interaction with the iterator
-        in the last `idle_timeout` seconds.
-        """
-        while self.active:
-            next_timeout = self._last_interaction_time + idle_timeout
-            await asyncio.sleep(next_timeout - time.monotonic())
-            if (
-                self._last_interaction_time + idle_timeout < time.monotonic()
-                and self.active
-            ):
-                # idle timeout has expired
-                await self._finish_with_error(
-                    IdleTimeout(
-                        (
-                            "Timed out waiting for next Row to be consumed. "
-                            f"(idle_timeout={idle_timeout:0.1f}s)"
-                        )
-                    )
-                )
-
-    def __aiter__(self):
-        """Implement the async iterator protocol."""
-        return self
-
-    async def __anext__(self) -> Row:
-        """
-        Implement the async iterator potocol.
-
-        Return the next item in the stream if active, or
-        raise an exception if the stream has been closed.
-        """
-        if self._error is not None:
-            raise self._error
+        # For each row
         try:
-            self._last_interaction_time = time.monotonic()
-            return await self._next_fn()
-        except Exception as e:
-            await self._finish_with_error(e)
-            raise e
+            while True:
+                c = await it.__anext__()
 
-    async def _finish_with_error(self, e: Exception):
-        """
-        Helper function to close the stream and clean up resources
-        after an error has occurred.
-        """
-        if self.active:
-            await self._merger.aclose()
-            self._error = e
-        if self._idle_timeout_task is not None:
-            self._idle_timeout_task.cancel()
-            self._idle_timeout_task = None
+                # EOS
+                if c is None:
+                    return
 
-    async def aclose(self):
-        """
-        Support closing the stream with an explicit call to aclose()
-        """
-        await self._finish_with_error(
-            StopAsyncIteration(f"{self.__class__.__name__} closed")
-        )
+                if c.reset_row:
+                    continue
+
+                row_key = c.row_key
+
+                if not row_key:
+                    raise InvalidChunk("first row chunk is missing key")
+
+                # Cells
+                cells = []
+
+                # shared per cell storage
+                family = c.family_name
+                qualifier = c.qualifier
+
+                try:
+                    # for each cell
+                    while True:
+                        f = c.family_name
+                        q = c.qualifier
+                        if f:
+                            family = f
+                            qualifier = q
+                        if q:
+                            qualifier = q
+
+                        ts = c.timestamp_micros
+                        labels = list(c.labels)
+                        value = c.value
+
+                        # merge split cells
+                        if c.value_size > 0:
+                            buffer = [value]
+                            # throws when early eos
+                            c = await it.__anext__()
+
+                            while c.value_size > 0:
+                                f = c.family_name
+                                q = c.qualifier
+                                t = c.timestamp_micros
+                                l = c.labels
+                                if f and f != family:
+                                    raise InvalidChunk("family changed mid cell")
+                                if q and q != qualifier:
+                                    raise InvalidChunk("qualifier changed mid cell")
+                                if t and t != ts:
+                                    raise InvalidChunk("timestamp changed mid cell")
+                                if l and l != labels:
+                                    raise InvalidChunk("labels changed mid cell")
+
+                                buffer.append(c.value)
+
+                                # throws when premature end
+                                c = await it.__anext__()
+
+                                if c.reset_row:
+                                    raise _ResetRow()
+                            else:
+                                buffer.append(c.value)
+                            value = b''.join(buffer)
+
+                        cells.append(Cell(row_key, family, qualifier, value, ts, labels))
+                        if c.commit_row:
+                            yield Row(row_key, cells)
+                            self._last_yielded_row_key = row_key
+                            if self._remaining_count is not None:
+                                self._remaining_count -= 1
+                            break
+                        c = await it.__anext__()
+                except _ResetRow:
+                    continue
+        except StopAsyncIteration:
+            pass
